split vm into separate crate
