+++ Cargo.toml
@@ -1,16 +1,11 @@
 [workspace]
 resolver = "3"
-members = ["opcode-gen", "ssa-builder", "wasm-runtime"]
+members = ["opcode-gen", "ssa-builder", "vm", "wasm-runtime"]
 
 [package]
 name = "mavros"
 version = "0.1.0"
 edition = "2024"
-
-[features]
-default = ["wasm", "plotting"]
-wasm = ["dep:inkwell", "dep:wasmtime"]
-plotting = ["dep:plotters"]
 
 [dependencies]
 fm = { git = "https://github.com/noir-lang/noir.git", branch = "master", package = "fm" }
@@ -38,16 +33,15 @@ ark-ff = { version = "0.5.0" }
 ark-serialize = { version = "0.5.0", features = ["derive"] }
 bincode = "1.3"
 
-# Optional heavy dependencies
-inkwell = { version = "0.5", features = ["llvm18-0"], optional = true }
-plotters = { version = "0.3.5", optional = true }
-wasmtime = { version = "31", optional = true }
+inkwell = { version = "0.5", features = ["llvm18-0"] }
+wasmtime = { version = "31" }
 
 tracing = "0.1.41"
 tracing-forest = "0.1.6"
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
 rand = "0.8"
 
+mavros-vm = { path = "vm" }
 opcode-gen = { path = "opcode-gen" }
 ssa-builder = { path = "ssa-builder" }
 cargo_metadata = "0.23.1"
