--- src/vm/mod.rs
@@ -1,3 +1,0 @@
-pub mod array;
-pub mod bytecode;
-pub mod interpreter;
