+++ vm/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "mavros-vm"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+ark-bn254 = { version = "0.5.0" }
+ark-ff = { version = "0.5.0" }
+noirc_abi = { git = "https://github.com/noir-lang/noir.git", branch = "master", package = "noirc_abi" }
+opcode-gen = { path = "../opcode-gen" }
+plotters = { version = "0.3.5" }
+serde = { version = "1.0.204", features = ["derive"] }
+tracing = "0.1.41"
